+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.0] - 2024-11-27
 ### Added
 - Gamepad Group Settings
 
@@ -70,7 +72,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add initial files
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/Audio/compare/3.1.0...main
+[Unreleased]: https://github.com/HyagoOliveira/Audio/compare/4.0.0...main
+[4.0.0]: https://github.com/HyagoOliveira/Audio/tree/4.0.0/
 [3.1.0]: https://github.com/HyagoOliveira/Audio/tree/3.1.0/
 [3.0.1]: https://github.com/HyagoOliveira/Audio/tree/3.0.1/
 [3.0.0]: https://github.com/HyagoOliveira/Audio/tree/3.0.0/